version 2.1.0 release

+++ src/main/java/cz/cvut/fel/aic/apdemo/DriveAgent.java
@@ -84,7 +84,7 @@ public class DriveAgent extends Agent implements Driver<PhysicalVehicle> {
 	}
 
 	@Override
-	public int getVelocity() {
+	public double getVelocity() {
 		return 15;
 	}
